requirements.txt
djangorestframework==3.1.3
Django==1.8.2
django-oauth-toolkit==0.8.1
django-cors-headers==1.1.0
django-nose==1.4
nose==1.3.7
httpie==0.9.2
ipython==3.1.0
factory-boy==2.5.2
